# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network Clone",
        "description": "Social Network Clone API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "feed/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get user images.",
                "operationId": "6964a8a48e86becab1fa0097e8a9870a",
                "responses": {
                    "200": {
                        "description": "User images",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/{post}/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get images for a post.",
                "operationId": "9ae6b3121d58b37a6bd0b2d8938763a5",
                "responses": {
                    "200": {
                        "description": "Post images",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Add an image to a post.",
                "operationId": "118c9c79d8debb9934c24a1188eea95e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image_url"
                                ],
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image added",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Post does not exist"
                    }
                }
            }
        },
        "feed/posts/{post}/images/{image}": {
            "put": {
                "tags": [
                    "Images"
                ],
                "summary": "Update an image for a post.",
                "operationId": "ebc7304f520169328625d6204dd52b58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "image_url"
                                ],
                                "properties": {
                                    "image_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image from a post.",
                "operationId": "4b9aceb3ea8d4adebdc9dde9c42a77f9",
                "responses": {
                    "200": {
                        "description": "Image deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image does not exist"
                    }
                }
            }
        },
        "feed/posts/{post}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get likes for a post.",
                "operationId": "a557c4ed9355284e28ac75e5abec4a5b",
                "responses": {
                    "200": {
                        "description": "Post likes",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like a post.",
                "operationId": "81f8c2b250bb5a5bbb63c963a7404299",
                "responses": {
                    "200": {
                        "description": "Post liked",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Unlike a post.",
                "operationId": "5dfdcac1e170c406e2caf9e82ddcda73",
                "responses": {
                    "200": {
                        "description": "Post unliked",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Get user likes.",
                "operationId": "ba9e50ee95128960ff4294a538c7bd06",
                "responses": {
                    "200": {
                        "description": "User likes",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
...
}